requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
flower==0.9.7
grpcio==1.35.0
grpcio-tools==1.35.0
humanize==3.2.0
importlib-metadata==3.4.0
kombu==4.6.11
prometheus-client==0.8.0
prompt-toolkit==3.0.14
protobuf==3.14.0
pytz==2021.1
redis==3.5.3
six==1.15.0
suds-py3==1.4.4.1
tornado==6.1
typing-extensions==3.7.4.3
vine==1.3.0
wcwidth==0.2.5
zipp==3.4.0
